requirements-dev.txt
autopep8
attrs >=17.4
bumpversion >=0.5
check-manifest >=0.36
doc8
flake8
invoke >=0.14
ipykernel
ipython >=5.8
isort
m2r
nbsphinx
pydocstyle
pytest >=3.2
sphinx_compas_theme >=0.4
sphinx ==2.4
twine
scikit-learn ==0.23.2


# install this package
-e .
